Add validation to copy tool parsing
Could keep the Copy cell idx in the struct, but then we gotta add
a check to prevent it goin off screen and becoming unrepresentable.

Could add a trait to make `all_transformations` easier, but it'd be
shaped kinda funky, and idk that it'd be totally helpful.

We'll see how it shakes out!!